+++ CHANGELOG.md
@@ -33,3 +33,7 @@ All notable autonomous changes should be recorded here.
 ## 2026-10-20T18:30:00.000000+00:00
 
 - Added a simple Python utility to generate a FreeCAD case design checklist and added a test.
+
+## 2026-11-01T14:00:00.000000+00:00
+
+- Added a simple Python utility to generate an ESP32/e-paper pinout helper and added a test.